apps/api/requirements.txt
requests
schedule
python-dotenv
openai
beautifulsoup4
google-genai
feedparser
yfinance
streamlit
plotly
python-json-logger
pytest
pytest-mock
psycopg2-binary
fastapi
sse-starlette
uvicorn
simhash
dateparser
pytz
pandas
GoogleNews
sentence-transformers
scikit-learn
# torch
socksio
akshare
pypinyin
redis
pandas_market_calendars
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
python-multipart
email-validator
sqlalchemy
resend
pyotp
qrcode[pil]
httpx>=0.24.0
webauthn>=2.0.0
sqlmodel
fastmcp
statsmodels
